+++ nicos_ess/devices/epics/__init__.py
@@ -1,4 +1,1 @@
-from nicos_ess.devices.epics.epics_devices import EpicsAnalogMoveable, \
-    EpicsDigitalMoveable, EpicsReadable, EpicsStringMoveable, \
-    EpicsStringReadable
-from nicos_ess.devices.epics.monitor_motor import EpicsMotor
+
