+++ nicos_ess/devices/epics/pva/__init__.py
@@ -1,0 +1,4 @@
+from nicos_ess.devices.epics.pva.epics_devices import EpicsAnalogMoveable, \
+    EpicsDigitalMoveable, EpicsReadable, EpicsStringMoveable, \
+    EpicsStringReadable
+from nicos_ess.devices.epics.pva.monitor_motor import EpicsMotor
