+++ nicos_ess/ymir/setups/motion_stages.py
@@ -1,7 +1,7 @@
 description = 'Stray motion stage(s) at Utgård/YMIR.'
 
 devices = dict(
-    m1=device('nicos_ess.devices.epics.EpicsMotor',
+    m1=device('nicos_ess.devices.epics.pva.EpicsMotor',
               epicstimeout=3.0,
               precision=0.1,
               description='Single axis positioner',
